fixed redundant journal logging
